chore: Rename `str.rs` to `from_str.rs`
